requirements.txt
astroid==2.2.5
backports.csv==1.0.7
colorama==0.4.1
defusedxml==0.6.0
diff-match-patch==20181111
Django==2.2
pip==19.2.3
django-crispy-forms==1.7.2
django-formtools==2.1
django-guardian==1.5.1
django-import-export==1.2.0
django-model-utils==3.2.0
django-notifications-latest==2.1.0
django-reversion==3.0.4
dwebsocket==0.5.11
et-xmlfile==1.0.1
future==0.15.2
httplib2==0.9.2
isort==4.3.21
jdcal==1.4.1
jsonfield==2.0.2
lazy-object-proxy==1.4.1
mccabe==0.6.1
mysqlclient==1.4.2.post1
odfpy==1.4.0
openpyxl==2.6.2
Pillow==6.0.0
Pillow-PIL==0.1.dev0
pylint==2.3.1
PyMySQL==0.9.3
pytz==2019.1
PyYAML==5.1.2
setuptools==40.8.0
six==1.12.0
sqlparse==0.3.0
swapper==1.1.1
tablib==0.13.0
typed-ast==1.4.0
wheel==0.33.4
wrapt==1.11.2
xlrd==1.2.0
xlwt==1.3.0
